+++ server/tests/db/sqlite3/schema.test.ts
@@ -154,6 +154,7 @@ describe("Gallery", () => {
     "title",
     "description",
     "icon",
+    "icon_source",
     "epoch",
     "epoch_type",
     "theme",
@@ -162,7 +163,7 @@ describe("Gallery", () => {
   ].join(",");
   test("Build create query", () =>
     expect(schema.gallery.buildCreateQuery()).toBe(
-      `INSERT INTO gallery (${cols}) VALUES (?,?,?,?,?,?,?,?,?)`
+      `INSERT INTO gallery (${cols}) VALUES (?,?,?,?,?,?,?,?,?,?)`
     ));
   test("Build select by id query", () =>
     expect(schema.gallery.buildSelectByIdQuery()).toBe(